requirements.txt
Django
djangorestframework
pytest-django
